Fixed an issue with the upgrade
